+++ apps/pie-storybook/stories/pie-modal.stories.ts
@@ -11,7 +11,9 @@ import {
 type ModalProps = ModalPropsBase & { slot: string }
 
 const defaultArgs: ModalProps = {
+    isDismissible: false,
     isOpen: true,
+    isFullWidthBelowMid: false,
     heading: 'Modal header',
     headingLevel: 'h2',
     size: 'medium',
@@ -22,6 +24,9 @@ export default {
     title: 'Modal',
     component: 'pie-modal',
     argTypes: {
+        isDismissible: {
+            control: 'boolean',
+        },
         isOpen: {
             control: 'boolean',
         },
@@ -31,6 +36,9 @@ export default {
         headingLevel: {
             control: 'select',
             options: headingLevels,
+        },
+        isFullWidthBelowMid: {
+            control: 'boolean',
         },
         size: {
             control: 'select',
@@ -60,7 +68,9 @@ const toggleModal = () => {
 };
 
 const Template = ({
+    isDismissible,
     isOpen,
+    isFullWidthBelowMid,
     heading,
     headingLevel,
     size,
@@ -68,9 +78,11 @@ const Template = ({
 }: ModalProps): TemplateResult => html`
         <pie-button @click=${toggleModal}>Toggle Modal</pie-button>
         <pie-modal
+        ?isDismissible="${isDismissible}"
         ?isOpen="${isOpen}"
         heading="${heading}"
         size="${size}"
+        ?isFullWidthBelowMid="${isFullWidthBelowMid}"
         headingLevel="${headingLevel}">
             ${slot}
         </pie-modal>
@@ -98,6 +110,7 @@ const PageContextTemplate = ({
     isOpen,
     heading,
     headingLevel,
+    isFullWidthBelowMid,
     slot,
 }: ModalProps) => html`
     <pie-button @click=${toggleModal}>Toggle Modal</pie-button>
@@ -105,6 +118,7 @@ const PageContextTemplate = ({
         ?isOpen="${isOpen}"
         heading="${heading}"
         headingLevel="${headingLevel}"
+        ?isFullWidthBelowMid="${isFullWidthBelowMid}"
     >
         ${slot}
     </pie-modal>