+++ packages/components/pie-modal/src/defs.ts
@@ -15,9 +15,26 @@ export interface ModalProps {
      */
     isOpen: boolean;
     /**
+     * When set to `true`:
+     *  1. The close button within the modal will be visible.
+     *  2. The user can dismiss the modal via the ESCAPE key, clicking the backdrop
+     *     or via a close button.
+     *
+     * When set to `false`:
+     *  1. The close button within the modal will be hidden.
+     *  2. The user can NOT dismiss the modal via the ESCAPE key or clicking the backdrop.
+     *
+     */
+    isDismissible: boolean;
+
+    /**
      * The size of the modal; this controls how wide it will appear on the page.
      */
     size: typeof sizes[number];
+    /**
+     * The width of the modal; this controls whether it will cover the full width of the page when below the mid breakpoint.
+     */
+    isFullWidthBelowMid: boolean;
 }
 
 /**